Rename HBD-AD to GSD-AD (#2573)
