+++ services/ui-src/src/measures/2025/GSDAD/data.ts
@@ -2,7 +2,7 @@ import { getCatQualLabels } from "../rateLabelText";
 import { MeasureTemplateData } from "shared/types/MeasureTemplate";
 import * as DC from "dataConstants";
 
-export const { categories, qualifiers } = getCatQualLabels("HBD-AD");
+export const { categories, qualifiers } = getCatQualLabels("GSD-AD");
 
 export const data: MeasureTemplateData = {
   type: "HEDIS",
