+++ services/ui-src/src/measures/2025/rateLabelText.ts
@@ -888,7 +888,7 @@ export const data = {
             }
         ]
     },
-    "HBD-AD": {
+    "GSD-AD": {
         "qualifiers": [
             {
                 "label": "Ages 18 to 64",
