+++ services/ui-src/src/measures/measureDescriptions.ts
@@ -428,7 +428,7 @@ export const measureDescriptions: MeasureList = {
     "FUM-AD":
       "Follow-Up After Emergency Department Visit for Mental Illness: Age 18 and Older",
     "FUM-HH": "Follow-Up After Emergency Department Visit for Mental Illness",
-    "HBD-AD": "Hemoglobin A1c Control for Patients with Diabetes",
+    "GSD-AD": "Glycemic Status Assessment for Patients with Diabetes",
     "HPCMI-AD":
       "Diabetes Care for People with Serious Mental Illness: Hemoglobin A1C (HbA1c) Poor Control (>9.0%)",
     "HVL-AD": "HIV Viral Load Suppression",
